pyproject.toml
[tool.poetry]
name = "cage-discord-bot"
version = "0.1.0"
description = ""
authors = ["Jake Brehm <jake.m.brehm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.3"
praw = "^6.5.1"
psycopg2-binary = "^2.8.5"
youtube_dl = "^2020.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
